try fixing deploy issue (#1)
