+++ src/components/Illustration/index.jsx
@@ -1,0 +1,140 @@
+export default function Illustration() {
+  return (
+    <svg
+      width="201"
+      height="249"
+      viewBox="0 0 201 249"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M125.361 41.8523C135.217 50.093 146.205 43.1937 156.736 42.6665C161.619 42.4221 166.493 43.5985 171.306 44.4761C175.875 45.3091 180.427 46.5207 184.603 48.5869C189.156 50.7601 193.028 54.1339 195.802 58.3444C196.353 59.2049 197.185 58.5695 197.178 57.5407C197.112 46.6167 192.385 47.5013 187.45 48.307C182.757 49.0729 178.768 44.4204 174.138 43.4191C169.293 42.445 164.396 41.7477 159.472 41.3307C154.133 40.8011 148.781 40.4418 143.415 40.2527C137.399 40.0407 131.381 40.0431 125.361 40.26C124.338 40.2964 124.333 41.8888 125.361 41.8523L125.361 41.8523Z"
+        fill="#3F3D56"
+      />
+      <path
+        d="M92.1159 129.733C92.2567 129.884 92.3333 130.084 92.3297 130.29C92.326 130.496 92.2424 130.693 92.0964 130.839C91.9504 130.985 91.7535 131.068 91.547 131.072C91.3406 131.076 91.1408 130.999 90.9898 130.859C84.4772 125.28 77.9647 119.706 71.4521 114.135C71.4446 114.126 71.4356 114.119 71.4256 114.113C70.9733 113.774 70.5455 113.404 70.1454 113.004C64.0843 112.298 57.4231 109.963 51.1124 108.949C45.5401 109.777 39.9837 110.765 34.3954 111.497C29.5265 112.173 24.618 112.524 19.7023 112.548C15.7024 112.527 16.8596 102.281 13.2315 100.392C9.45997 98.4279 1.8964 104.783 0.929605 100.813C0.451934 98.7807 0.141034 96.7127 0 94.6299V81.6849C0.515267 76.0484 9.49713 77.7872 13.7627 73.8544C20.1637 67.9525 22.3371 54.1074 30.5282 53.9429C40.2897 53.7371 50.0542 54.1644 59.7603 55.2219C67.6752 56.0924 75.9992 56.9363 83.3192 60.3065C86.4586 61.7554 89.4174 70.1359 91.5794 72.8586C94.0468 76.0392 95.4433 79.9185 95.5687 83.9406C95.9565 92.4803 92.0681 100.293 89.938 108.355C88.0894 115.334 87.0483 123.794 92.1159 129.733Z"
+        fill="#F2F2F2"
+      />
+      <path
+        d="M72.1923 114.953C78.7949 120.607 85.3984 126.26 92.0029 131.913C92.154 132.054 92.3539 132.13 92.5605 132.127C92.7671 132.123 92.9642 132.039 93.1103 131.893C93.2564 131.747 93.3401 131.55 93.3437 131.344C93.3474 131.138 93.2706 130.938 93.1297 130.787C88.0624 124.852 89.1036 116.392 90.9508 109.409C93.0836 101.346 96.9718 93.5361 96.5826 84.9985C96.4592 80.9753 95.0621 77.0948 92.5922 73.915C90.3595 71.2393 87.5315 69.1222 84.3341 67.7329C77.0145 64.3626 68.6921 63.515 60.7736 62.6468C51.0687 61.5883 41.3053 61.1604 31.5449 61.3659C23.3514 61.5342 13.7667 61.6837 7.36611 67.5873C1.53703 72.9638 0.740868 81.4187 0.79708 88.8835C0.829642 93.2083 0.912688 97.6505 1.94169 101.871C2.90899 105.839 5.00664 109.703 8.77938 111.667C12.4082 113.555 16.7194 113.584 20.7156 113.604C25.6312 113.583 30.5398 113.231 35.408 112.55C45.1513 111.277 54.8017 109.22 64.6239 108.598C67.0274 108.446 69.4382 108.379 71.8464 108.429C72.8723 108.45 72.8722 106.858 71.8464 106.837C62.2961 106.637 52.844 108.276 43.4474 109.768C34.3587 111.212 25.0677 112.572 15.8424 111.842C13.82 111.747 11.8315 111.288 9.97301 110.485C8.1781 109.633 6.65937 108.294 5.59042 106.62C3.18872 102.883 2.73367 98.2014 2.52621 93.8417C2.28954 90.0686 2.3246 86.2833 2.6311 82.5152C2.98505 78.8617 3.83746 75.1558 5.7846 72.0017C10.2261 64.8069 19.2063 61.813 26.8705 63.1055C35.5011 64.561 42.1029 71.4538 55.2727 63.7078C59.4068 61.2763 64.8241 64.66 69.5677 65.375C73.1128 65.8634 76.6095 66.6542 80.0196 67.7385C83.3007 68.739 86.3577 70.3625 89.0232 72.5202C91.8131 74.8776 93.7552 78.0815 94.5538 81.6439C96.383 89.8477 92.7573 98.0363 90.3614 105.714C88.2257 112.557 86.4037 120.594 89.2794 127.5C89.95 129.104 90.869 130.594 92.0029 131.913L93.1297 130.787C86.5271 125.132 79.9235 119.479 73.3191 113.827C72.5443 113.163 71.412 114.284 72.1922 114.953H72.1923Z"
+        fill="#E6E6E6"
+      />
+      <path
+        d="M44.0582 77.9068H20.0368C19.8164 77.9068 19.605 77.8193 19.4492 77.6636C19.2933 77.5079 19.2057 77.2967 19.2057 77.0765C19.2057 76.8562 19.2933 76.645 19.4492 76.4893C19.605 76.3336 19.8164 76.2461 20.0368 76.2461H44.0582C44.2786 76.2461 44.49 76.3336 44.6458 76.4893C44.8017 76.645 44.8893 76.8562 44.8893 77.0765C44.8893 77.2967 44.8017 77.5079 44.6458 77.6636C44.49 77.8193 44.2786 77.9068 44.0582 77.9068Z"
+        fill="white"
+      />
+      <path
+        d="M44.0582 95.6144H20.0368C19.9277 95.6144 19.8196 95.5929 19.7188 95.5512C19.618 95.5095 19.5263 95.4483 19.4492 95.3712C19.372 95.2941 19.3108 95.2026 19.269 95.1018C19.2272 95.0011 19.2057 94.8931 19.2057 94.7841C19.2057 94.675 19.2272 94.567 19.269 94.4663C19.3108 94.3655 19.372 94.274 19.4492 94.1969C19.5263 94.1198 19.618 94.0586 19.7188 94.0169C19.8196 93.9752 19.9277 93.9537 20.0368 93.9537H44.0582C44.2786 93.9537 44.49 94.0412 44.6458 94.1969C44.8017 94.3526 44.8893 94.5638 44.8893 94.7841C44.8893 95.0043 44.8017 95.2155 44.6458 95.3712C44.49 95.5269 44.2786 95.6144 44.0582 95.6144Z"
+        fill="white"
+      />
+      <path
+        d="M63.5673 86.7664H20.0368C19.8164 86.7664 19.605 86.6789 19.4492 86.5232C19.2933 86.3675 19.2057 86.1563 19.2057 85.9361C19.2057 85.7158 19.2933 85.5046 19.4492 85.3489C19.605 85.1932 19.8164 85.1057 20.0368 85.1057H63.5673C63.7877 85.1057 63.9991 85.1932 64.155 85.3489C64.3108 85.5046 64.3984 85.7158 64.3984 85.9361C64.3984 86.1563 64.3108 86.3675 64.155 86.5232C63.9991 86.6789 63.7877 86.7664 63.5673 86.7664Z"
+        fill="white"
+      />
+      <path
+        d="M118.676 170.07C118.817 170.221 118.894 170.421 118.89 170.627C118.886 170.833 118.803 171.03 118.657 171.176C118.511 171.322 118.314 171.405 118.107 171.409C117.901 171.412 117.701 171.336 117.55 171.195C111.037 165.617 104.525 160.042 98.0123 154.471C98.0048 154.463 97.9958 154.456 97.9858 154.45C97.5335 154.111 97.1057 153.74 96.7056 153.341C90.6445 152.635 83.9833 150.3 77.6726 149.286C72.1003 150.114 65.4815 156.409 59.8932 157.141C55.0243 157.817 50.1158 158.168 45.2001 158.192C41.2002 158.171 42.3574 145.272 38.7293 143.382C34.9578 141.418 28.4566 145.12 27.4898 141.15C26.9958 139.128 24.5566 137.052 24.3973 134.967V122.022C24.9125 116.385 28.6478 110.802 32.9134 106.869C39.3144 100.967 48.8973 100.813 57.0884 100.648C66.8499 100.443 76.6144 100.87 86.3205 101.928C94.2355 102.798 102.559 103.642 109.879 107.012C113.077 108.403 115.905 110.52 118.14 113.195C120.607 116.376 122.004 120.255 122.129 124.277C122.517 132.817 118.628 140.63 116.498 148.692C114.65 155.671 113.608 164.131 118.676 170.07Z"
+        fill="#F2F2F2"
+      />
+      <path
+        d="M98.7524 155.289C105.355 160.944 111.959 166.597 118.563 172.249C118.714 172.39 118.914 172.467 119.121 172.463C119.327 172.46 119.524 172.376 119.67 172.23C119.817 172.084 119.9 171.887 119.904 171.681C119.907 171.474 119.831 171.275 119.69 171.124C114.622 165.189 115.664 156.729 117.511 149.745C119.644 141.683 123.532 133.873 123.143 125.335C123.019 121.312 121.622 117.432 119.152 114.252C116.92 111.576 114.092 109.459 110.894 108.07C103.575 104.699 95.2522 103.852 87.3337 102.983C77.6288 101.925 67.8654 101.497 58.1051 101.703C49.9115 101.871 40.3268 102.02 33.9262 107.924C28.0972 113.3 27.301 121.755 27.3572 129.22C27.3898 133.545 27.4728 137.987 28.5018 142.208C29.4691 146.176 31.5668 150.04 35.3395 152.003C38.9683 153.892 43.2795 153.921 47.2757 153.94C52.1913 153.92 57.0999 153.568 61.9681 152.887C71.7114 151.613 81.3618 149.557 91.184 148.935C93.5876 148.783 95.9983 148.715 98.4065 148.766C99.4324 148.787 99.4323 147.195 98.4065 147.173C88.8562 146.974 79.4041 148.613 70.0076 150.105C60.9189 151.548 51.6278 152.909 42.4025 152.179C40.3801 152.084 38.3916 151.624 36.5331 150.822C34.7382 149.97 33.2195 148.63 32.1505 146.956C29.7488 143.219 29.2938 138.538 29.0863 134.178C28.8497 130.405 28.8847 126.62 29.1912 122.852C29.5452 119.198 30.3976 115.493 32.3447 112.338C36.7863 105.144 46.9122 99.2122 53.4306 103.442C60.8781 108.275 70.403 108.484 81.8328 104.044C86.304 102.308 91.3842 104.997 96.1279 105.712C99.6729 106.2 103.17 106.991 106.58 108.075C109.861 109.076 112.918 110.699 115.583 112.857C118.373 115.214 120.315 118.418 121.114 121.981C122.943 130.184 119.317 138.373 116.922 146.05C114.786 152.894 112.964 160.931 115.839 167.836C116.51 169.441 117.429 170.93 118.563 172.249L119.69 171.124C113.087 165.469 106.484 159.816 99.8792 154.163C99.1044 153.5 97.9721 154.621 98.7524 155.289H98.7524Z"
+        fill="#E6E6E6"
+      />
+      <path
+        d="M70.6183 118.244H46.597C46.3766 118.244 46.1652 118.156 46.0093 118C45.8535 117.845 45.7659 117.633 45.7659 117.413C45.7659 117.193 45.8535 116.982 46.0093 116.826C46.1652 116.67 46.3766 116.583 46.597 116.583H70.6183C70.8388 116.583 71.0502 116.67 71.206 116.826C71.3619 116.982 71.4494 117.193 71.4494 117.413C71.4494 117.633 71.3619 117.845 71.206 118C71.0502 118.156 70.8388 118.244 70.6183 118.244Z"
+        fill="white"
+      />
+      <path
+        d="M70.6183 135.951H46.597C46.3766 135.951 46.1652 135.864 46.0093 135.708C45.8535 135.552 45.7659 135.341 45.7659 135.121C45.7659 134.901 45.8535 134.689 46.0093 134.534C46.1652 134.378 46.3766 134.29 46.597 134.29H70.6183C70.8388 134.29 71.0502 134.378 71.206 134.534C71.3619 134.689 71.4494 134.901 71.4494 135.121C71.4494 135.341 71.3619 135.552 71.206 135.708C71.0502 135.864 70.8388 135.951 70.6183 135.951Z"
+        fill="white"
+      />
+      <path
+        d="M90.1275 127.103H46.597C46.3766 127.103 46.1652 127.016 46.0093 126.86C45.8535 126.704 45.7659 126.493 45.7659 126.273C45.7659 126.053 45.8535 125.841 46.0093 125.686C46.1652 125.53 46.3766 125.442 46.597 125.442H90.1275C90.3479 125.442 90.5593 125.53 90.7151 125.686C90.871 125.841 90.9586 126.053 90.9586 126.273C90.9586 126.493 90.871 126.704 90.7151 126.86C90.5593 127.016 90.3479 127.103 90.1275 127.103Z"
+        fill="white"
+      />
+      <path
+        d="M144.593 221.668L148.407 225.254L164.375 211.936L158.746 206.643L144.593 221.668Z"
+        fill="#FFB6B6"
+      />
+      <path
+        d="M151.537 237.175L149.418 227.302L151.986 224.575L150.604 223.419L146.31 219.845L145.26 218.966L135.385 227.841L138.247 230.529L141.299 228.312L140.061 232.238L147.642 239.369C148.006 239.71 148.469 239.927 148.965 239.987C149.46 240.047 149.962 239.948 150.397 239.703C150.832 239.458 151.177 239.081 151.382 238.626C151.587 238.171 151.641 237.663 151.537 237.175V237.175Z"
+        fill="#2F2E41"
+      />
+      <path
+        d="M126.002 235.536L131.18 236.322L136.747 216.298L129.105 215.137L126.002 235.536Z"
+        fill="#FFB6B6"
+      />
+      <path
+        d="M140.535 244.355L133.176 237.433L133.739 233.731L131.944 233.565L126.379 233.061L125.015 232.934L121.935 245.843L125.819 246.431L127.069 242.873L128.283 246.807L138.577 248.371C139.071 248.446 139.575 248.361 140.017 248.129C140.459 247.897 140.815 247.53 141.034 247.082C141.252 246.634 141.322 246.127 141.232 245.637C141.142 245.146 140.898 244.697 140.535 244.355L140.535 244.355Z"
+        fill="#2F2E41"
+      />
+      <path
+        d="M175.813 93.286L190.661 119.857C190.661 119.857 190.042 174.505 192.015 176.532C193.989 178.559 195.775 180.377 193.989 180.361C192.202 180.344 192.778 193.272 192.778 193.272L159.844 218.748L151.876 212.379L171.865 187.726L160.826 137.265C160.826 137.265 152.895 184.986 151.465 186.414C150.036 187.842 149.174 187.832 150.677 189.265C152.18 190.698 150.677 193.94 150.677 193.94L135.008 226.432L126.508 225.361C126.508 225.361 127.982 198.835 133.113 187.448V133.633C132.726 138.247 150.804 91.547 139.749 95.0521C143.247 93.9432 175.813 93.286 175.813 93.286Z"
+        fill="#2F2E41"
+      />
+      <path
+        d="M169.001 29.9637H156.726L154.988 36.9095L145.55 39.5171C145.55 39.5171 140.881 70.449 144.802 72.944C148.723 75.439 145.989 83.2435 145.989 83.2435C145.989 83.2435 136.225 93.7796 137.397 95.3405C138.568 96.9014 140.945 94.5263 138.195 97.2748C135.444 100.023 133.881 99.633 134.663 100.413C135.146 100.908 135.52 101.499 135.762 102.146L132.393 110.91C132.393 110.91 150.285 114.462 159.268 109.389C168.251 104.316 188.17 114.071 188.17 114.071C188.17 114.071 187.145 108.949 186.29 108.193C185.436 107.437 183.443 109.349 184.83 106.832C186.217 104.316 187.388 105.486 186.217 104.316C185.045 103.145 179.968 87.536 179.968 87.536C179.968 87.536 176.062 78.951 180.358 75.439C184.655 71.927 186.29 60.4076 186.29 60.4076L181.14 39.5171L172.547 37.5871L169.001 29.9637Z"
+        fill="#E6E6E6"
+      />
+      <path
+        d="M120.409 109.717L123.358 97.9098L116.842 95.286L114.17 107.88C113.56 108.331 113.088 108.942 112.807 109.646C112.526 110.35 112.448 111.119 112.58 111.865C112.712 112.611 113.05 113.306 113.556 113.87C114.062 114.435 114.715 114.848 115.442 115.062C116.17 115.276 116.943 115.283 117.674 115.083C118.406 114.882 119.067 114.482 119.583 113.926C120.099 113.371 120.449 112.683 120.596 111.939C120.742 111.195 120.677 110.426 120.409 109.717V109.717Z"
+        fill="#FFB6B6"
+      />
+      <path
+        d="M149.048 38.7279L148.853 38.6944C147.044 38.3839 145.187 38.5186 143.442 39.0868C141.697 39.655 140.117 40.6397 138.838 41.9555C135.501 45.3862 120.603 70.8996 121.139 75.1636C121.639 79.1401 121.831 78.2556 120.011 80.2859C118.191 82.3162 117.499 79.2243 117.471 83.4514C117.443 87.6785 119.059 87.6417 117.341 88.6753C115.623 89.7089 114.487 88.9889 114.801 91.8408C115.114 94.6927 114.634 95.4495 114.532 96.4463C114.431 97.4431 111.964 103.839 111.964 103.839C111.964 103.839 119.224 105.787 122.152 108.704C122.152 108.704 122.485 101.486 123.963 100.831C125.441 100.176 126.818 100.517 126.402 98.6623C125.987 96.8071 123.853 95.9855 126.467 96.0503C129.081 96.1151 133.119 82.2251 133.119 82.2251L150.579 50.3969L149.048 38.7279Z"
+        fill="#E6E6E6"
+      />
+      <path
+        d="M188.45 115.566L188.975 103.408L195.968 102.738L194.955 115.573C195.412 116.178 195.691 116.898 195.76 117.653C195.83 118.407 195.687 119.166 195.348 119.844C195.01 120.522 194.488 121.093 193.843 121.491C193.198 121.889 192.454 122.1 191.696 122.099C190.937 122.098 190.194 121.886 189.55 121.486C188.905 121.087 188.385 120.515 188.048 119.837C187.711 119.158 187.57 118.398 187.641 117.644C187.712 116.889 187.992 116.17 188.45 115.566Z"
+        fill="#FFB6B6"
+      />
+      <path
+        d="M181.14 39.3762L181.336 39.3992C183.159 39.6141 184.902 40.2696 186.414 41.3091C187.926 42.3485 189.162 43.7405 190.014 45.3646C192.241 49.6002 199.285 78.2879 197.56 82.2248C195.951 85.8963 196.018 84.9935 197.187 87.4562C198.356 89.9189 199.898 87.1502 198.724 91.2117C197.551 95.2731 196.012 94.7797 197.366 96.2578C198.72 97.7358 200.014 97.3674 198.903 100.013C197.793 102.659 198.039 103.521 197.853 104.506C197.668 105.49 198.218 112.323 198.218 112.323C198.218 112.323 190.702 112.133 187.067 114.1C187.067 114.1 188.796 107.085 187.565 106.038C186.333 104.99 184.917 104.928 185.842 103.267C186.767 101.606 189.046 101.422 186.521 100.744C183.996 100.065 184.067 85.6011 184.067 85.6011L176.359 50.1321L181.14 39.3762Z"
+        fill="#E6E6E6"
+      />
+      <path
+        d="M134.972 21.7687C134.875 25.9551 136.44 30.0087 138.69 33.5692C139.441 31.5018 140.494 29.5564 141.813 27.7955C141.181 31.6967 141.854 35.6972 143.727 39.1778C143.794 39.225 143.862 39.2689 143.932 39.3092C145.766 40.4112 148.091 40.8576 150.074 40.0616C145.976 35.6477 145.454 28.3344 148.89 23.3834C149.863 21.9786 151.121 20.7361 151.724 19.1374C152.675 16.5865 151.761 13.7779 151.184 11.1183C150.644 8.64307 150.55 5.6975 152.212 3.88345C152.034 3.62994 151.793 3.42814 151.511 3.29874C151.23 3.16934 150.919 3.11699 150.611 3.14705C149.586 3.29168 148.84 4.18162 148.365 5.09821C147.89 6.01477 147.577 7.03066 146.915 7.82231C145.689 9.30412 143.601 9.65444 141.841 10.4354C137.58 12.3318 135.074 17.1103 134.972 21.7687Z"
+        fill="#2F2E41"
+      />
+      <path
+        d="M170.842 19.375C170.598 18.5235 172.277 23.7289 171.467 23.8229C167.778 24.2457 162.714 26.4469 160.145 27.4738C160.029 27.4604 159.913 27.4469 159.801 27.4302C159.526 27.4 159.257 27.3563 158.991 27.306C153.009 26.1751 150.708 16.9932 150.708 16.9093C150.708 16.4965 151.7 5.15943 153.262 4.99164C153.672 4.94716 154.084 4.92475 154.497 4.92452H158.373C159.175 4.75183 159.988 4.6341 160.806 4.57215H160.813C168.616 3.97815 174.769 8.7301 174.194 14.5123C174.19 14.519 174.183 14.5257 174.18 14.5325C174.013 14.7808 173.867 15.0157 173.738 15.2406C173.612 15.452 173.506 15.6533 173.414 15.8446C172.703 17.3212 174.531 19.1792 173.319 19.5261C171.707 19.9873 177.559 35.1478 171.178 35.1478C174.282 36.0296 171.514 21.7242 170.842 19.375Z"
+        fill="#2F2E41"
+      />
+      <path
+        d="M163.876 28.2706C168.944 28.2706 173.053 24.1654 173.053 19.1015C173.053 14.0376 168.944 9.93246 163.876 9.93246C158.808 9.93246 154.699 14.0376 154.699 19.1015C154.699 24.1654 158.808 28.2706 163.876 28.2706Z"
+        fill="#FFB6B6"
+      />
+      <path
+        d="M174.194 15.0124C174.04 15.0929 173.891 15.1667 173.738 15.2406C173.578 15.3178 173.418 15.395 173.258 15.4654C170.185 16.8581 167.684 17.0125 166.166 15.0124C166.125 13.96 165.889 12.9243 165.468 11.9585C165.543 13.0378 165.258 14.1114 164.658 15.0124H161.602C158.455 17.0997 156.838 17.3179 159.023 24.1706C159.632 26.0801 162.054 37.1915 161.101 38.8393C160.826 38.8091 157.893 26.6387 157.628 26.5883C151.645 25.4574 147.844 15.6097 148.331 15.0527C148.328 13.4685 148.693 11.9052 149.396 10.4853C150.285 8.68626 151.672 7.17935 153.391 6.14273C154.416 5.6462 155.484 5.24515 156.583 4.94468C156.61 4.93796 156.634 4.93123 156.661 4.92454C156.998 4.83392 157.342 4.75002 157.686 4.67955C158.396 4.52975 159.114 4.42107 159.836 4.35403C159.887 4.35068 159.938 4.34732 159.986 4.34732C160.275 4.348 160.558 4.42561 160.806 4.57217C160.81 4.57217 160.81 4.57551 160.813 4.57551C160.991 4.67918 161.146 4.81724 161.269 4.98158C161.486 5.2581 161.604 5.59985 161.602 5.95145H165.005C165.131 5.95145 165.257 5.9548 165.383 5.9615C167.654 6.03905 169.815 6.96062 171.442 8.54585C173.069 10.1311 174.045 12.2659 174.18 14.5325C174.19 14.6935 174.194 14.8513 174.194 15.0124Z"
+        fill="#2F2E41"
+      />
+      <path
+        d="M154.129 12.0813C157.468 12.0813 160.174 9.37677 160.174 6.04063C160.174 2.70448 157.468 0 154.129 0C150.79 0 148.083 2.70448 148.083 6.04063C148.083 9.37677 150.79 12.0813 154.129 12.0813Z"
+        fill="#2F2E41"
+      />
+      <path
+        d="M190.357 112.752C191.096 113.531 189.976 114.543 189.114 113.875C181.924 108.306 174.734 102.74 167.545 97.1778C167.536 97.169 167.526 97.1619 167.515 97.1566C167.019 96.8204 166.547 96.4504 166.102 96.0491C159.411 95.3443 152.057 93.0128 145.09 92.0007C138.938 92.8273 132.804 93.8129 126.634 94.5442C121.254 95.2197 115.837 95.57 110.414 95.5934C105.998 95.5722 101.242 95.5457 97.2361 93.6593C93.0724 91.6986 90.756 87.8409 89.6886 83.8773C89.1623 81.8543 88.8188 79.7882 88.6624 77.7039V64.7796C89.2312 59.152 90.9671 53.5775 95.6762 49.6509C102.743 43.7584 113.322 43.6047 122.365 43.4405C133.138 43.2349 143.914 43.6613 154.637 44.7175C163.375 45.5866 172.565 46.4291 180.646 49.794C184.112 51.2406 187.378 53.249 189.765 55.9674C192.451 59.047 194.005 62.9502 194.169 67.0317C194.597 75.5578 190.304 83.358 187.953 91.4072C185.912 98.3754 184.763 106.822 190.357 112.752Z"
+        fill="#967C6E"
+      />
+      <path
+        d="M137.019 57.7385H112.997C112.888 57.7385 112.78 57.717 112.679 57.6753C112.578 57.6336 112.487 57.5724 112.41 57.4953C112.332 57.4182 112.271 57.3267 112.23 57.2259C112.188 57.1252 112.166 57.0172 112.166 56.9081C112.166 56.7991 112.188 56.6911 112.23 56.5904C112.271 56.4896 112.332 56.3981 112.41 56.321C112.487 56.2439 112.578 56.1827 112.679 56.141C112.78 56.0992 112.888 56.0778 112.997 56.0778H137.019C137.239 56.0778 137.451 56.1652 137.606 56.321C137.762 56.4767 137.85 56.6879 137.85 56.9081C137.85 57.1284 137.762 57.3396 137.606 57.4953C137.451 57.651 137.239 57.7385 137.019 57.7385Z"
+        fill="white"
+      />
+      <path
+        d="M137.019 75.4461H112.997C112.777 75.4461 112.566 75.3586 112.41 75.2029C112.254 75.0472 112.166 74.8359 112.166 74.6157C112.166 74.3955 112.254 74.1843 112.41 74.0285C112.566 73.8728 112.777 73.7853 112.997 73.7853H137.019C137.239 73.7853 137.451 73.8728 137.606 74.0285C137.762 74.1843 137.85 74.3955 137.85 74.6157C137.85 74.8359 137.762 75.0472 137.606 75.2029C137.451 75.3586 137.239 75.4461 137.019 75.4461Z"
+        fill="white"
+      />
+      <path
+        d="M156.528 66.5981H112.997C112.777 66.5981 112.566 66.5106 112.41 66.3549C112.254 66.1992 112.166 65.9879 112.166 65.7677C112.166 65.5475 112.254 65.3363 112.41 65.1805C112.566 65.0248 112.777 64.9373 112.997 64.9373H156.528C156.748 64.9373 156.96 65.0248 157.115 65.1805C157.271 65.3363 157.359 65.5475 157.359 65.7677C157.359 65.9879 157.271 66.1992 157.115 66.3549C156.96 66.5106 156.748 66.5981 156.528 66.5981Z"
+        fill="white"
+      />
+      <path
+        d="M195.803 58.3444C198.029 61.894 199.289 65.9795 198.903 70.1907C198.57 73.8243 197.078 77.2416 195.447 80.462C193.757 83.7992 191.865 87.0399 190.483 90.5243C188.966 94.3223 188.405 98.4344 188.85 102.499C189.123 105.139 189.656 107.742 190.158 110.345L191.691 118.291L193.023 117.516L170.99 98.4844C170.218 97.8173 169.086 98.9388 169.863 99.6103L191.896 118.642C192.022 118.76 192.184 118.835 192.356 118.855C192.528 118.874 192.701 118.838 192.851 118.751C193.001 118.664 193.118 118.531 193.186 118.371C193.254 118.212 193.269 118.035 193.228 117.867C192.327 113.199 191.263 108.528 190.604 103.819C190.225 101.444 190.171 99.0288 190.446 96.6395C190.715 94.8184 191.194 93.0344 191.873 91.323C194.516 84.4236 199.465 78.3881 200.42 70.887C201.017 66.1872 199.673 61.5169 197.179 57.5408C196.635 56.6745 195.256 57.4728 195.803 58.3444V58.3444Z"
+        fill="#3F3D56"
+      />
+      <path
+        d="M172.383 91.5188L139.798 95.3704L123.505 97.2962C118.542 97.8829 113.369 98.7067 108.409 97.6685C100.06 95.9208 94.8752 88.3614 93.2442 80.3947C92.2346 75.4633 92.592 70.3035 93.4414 65.3768C93.8942 62.7502 94.4918 60.151 95.1376 57.5659C95.7131 55.262 96.1424 52.7811 97.1369 50.6124C97.9304 48.849 99.1112 47.2868 100.592 46.0417C102.072 44.7966 103.814 43.9005 105.689 43.42C106.859 43.1803 108.046 43.0233 109.238 42.9501C110.579 42.8157 111.92 42.6921 113.262 42.5791C115.946 42.3531 118.634 42.1701 121.324 42.0301C122.669 41.9601 124.015 41.9008 125.361 41.8523C126.383 41.8149 126.388 40.2225 125.361 40.2601C119.76 40.4649 114.169 40.8528 108.59 41.4238C106.317 41.6578 104.216 41.9959 102.164 43.0597C100.391 43.9862 98.8336 45.2764 97.5943 46.8459C96.3244 48.465 95.3928 50.3225 94.8549 52.3081C94.1216 54.9465 93.456 57.611 92.8488 60.281C91.6249 65.663 90.7451 71.2103 91.1538 76.7433C91.7583 84.9269 95.6796 93.6579 103.29 97.554C107.784 99.8547 112.912 99.9578 117.836 99.5185C123.328 99.0286 128.804 98.2621 134.279 97.6149L168.178 93.608L172.383 93.111C173.389 92.9921 173.401 91.3984 172.383 91.5188H172.383Z"
+        fill="#3F3D56"
+      />
+    </svg>
+  );
+}
